requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
psycopg2==2.9.6
pydantic==1.10.9
sniffio==1.3.0
SQLAlchemy~=2.0.18
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
typing_extensions==4.6.3
uvicorn==0.22.0
yoyo-migrations==8.2.0
zipp==3.15.0
pytest~=7.3.2
pip~=23.1.2
wheel~=0.40.0
cryptography~=41.0.1
Pillow~=9.5.0
Jinja2~=3.1.2
requests~=2.31.0
websockets~=11.0.3
httptools~=0.5.0
watchfiles~=0.19.0
mypy~=1.3.0
itsdangerous~=2.1.2
PyYAML~=6.0
setuptools~=65.5.0
packaging~=23.1
pyparsing~=3.1.0
PyJWT==2.7.0
python-multipart~=0.0.6
bs4~=0.0.1
beautifulsoup4~=4.12.2
lxml~=4.9.2
soupsieve~=2.4.1
cffi~=1.15.1
MarkupSafe~=2.1.3
pluggy~=1.2.0
iniconfig~=2.0.0
numpy~=1.24.3
certifi~=2023.5.7
urllib3~=2.0.3
pycparser~=2.21
python-escpos==2.2.0
pandas==2.1.0
tqdm~=4.65.0
openpyxl==3.1.2
matplotlib~=3.7.1
openai~=0.28.1
XlsxWriter==3.1.5
transliterate==1.10.2
xlrd==2.0.1